backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
polars==0.20.6
pyarrow==14.0.1
pillow==10.1.0
tqdm==4.66.1
psutil==5.9.6
datasets==2.15.0
huggingface-hub==0.19.4 